+++ testsuite/services_impl/api/api_container_initializer.go
@@ -48,8 +48,8 @@ func (initializer ApiContainerInitializer) GetUsedPorts() map[string]bool {
 	}
 }
 
-func (initializer ApiContainerInitializer) GetServiceFromIp(ipAddr string) services.Service {
-	return NewApiService(ipAddr, port)
+func (initializer ApiContainerInitializer) GetService(serviceId services.ServiceID, ipAddr string) services.Service {
+	return NewApiService(serviceId, ipAddr, port)
 }
 
 func (initializer ApiContainerInitializer) GetFilesToMount() map[string]bool {